Following recommended webpack way of excluding devtool code from production build
